+++ GridGame/Assets/_Scripts/OddsMaker.cs
@@ -44,12 +44,12 @@ public class OddsMaker : MonoBehaviour
 
             // Add attraction for positive types
             foreach (var p in relation.positiveTypes)
-                if (agg.TryGetValue(p.type, out int q))
+                if (agg.TryGetValue(p.type, out float q))
                     dirWeights[i] += p.weight * q;
 
             // Subtract repulsion for negative types
             foreach (var n in relation.negativeTypes)
-                if (agg.TryGetValue(n.type, out int q2))
+                if (agg.TryGetValue(n.type, out float q2))
                     dirWeights[i] -= n.weight * q2;
 
             dirWeights[i] += neighborContainers[i].GetManaGap() * 2;
